requirements.txt
pymongo==4.13.2
pytest==8.4.1
Werkzeug==3.1.3
pydantic~=2.11.7
Flask~=3.1.1
python-dotenv~=1.1.1